backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
openai>=1.10.0
spotipy>=2.23.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
httpx>=0.26.0
python-multipart>=0.0.6
ollama>=0.5.0
